oops, require numeric instead of integer columns
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/dplr@640 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ branches/zero-is-missing/R/helpers.R
@@ -22,7 +22,7 @@ trim.rwl.df <- function(rwl, zero.is.na=TRUE) {
 
 ### Creates a rwl data.frame with consecutive years
 complete.rwl.df <- function(rwl, as.mat=FALSE) {
-    if (!all(vapply(rwl, is.vector, TRUE, mode="integer"))) {
+    if (!all(vapply(rwl, is.vector, TRUE, mode="numeric"))) {
         stop("'rwl' must not have columns containing a matrix")
     }
     cnames <- names(rwl)
